#44-rename module types.h
Done.
Fixed typo in separation rules in module `mrgrand49507.py`.
