+++ PyRandLib/mrgrand49507.py
@@ -20,11 +20,11 @@ OUT  OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-#----------------------------------------------------------------------------=
+#=============================================================================
 from .basemrg import BaseMRG
 
 
-#----------------------------------------------------------------------------=
+#=============================================================================
 class MRGRand49507( BaseMRG ):
     """
     Pseudo-random numbers generator  - Definition of a fast 31-bits Multiple Recursive 
@@ -97,13 +97,13 @@ class MRGRand49507( BaseMRG ):
     should definitively pass.
     """    
     
-    #------------------------------------------------------------------------=
+    #=========================================================================
     # 'protected' constant
     _LIST_SIZE = 1597       # this 'DX-1597-2-7' MRG is based on a suite containing 1597 integers
     _MODULO    = 2_147_483_647 # i.e. 0x7fffffff, or (1<<31)-1, the modulo for DX-1597-2-7 MRG
             
  
-    #------------------------------------------------------------------------=
+    #=========================================================================
     def random(self) -> float:
         """This is the core of the pseudo-random generator.
         
@@ -124,4 +124,4 @@ class MRGRand49507( BaseMRG ):
         # then returns float value within [0.0, 1.0)
         return  myValue * 4.656612875245797e-10  # / 2_147_483_647.0
  
-#----=   end of module   mrgrand49507.py   ----------------------------------=
+#=====   end of module   mrgrand49507.py   ===================================
